part1-database-etl/requirements.txt
pandas==2.1.4
mysql-connector-python==9.0.0
numpy==1.26.4
python-dotenv==1.0.0
